particle.lhcb restructured/simplified
