+++ src/particle/lhcb/__init__.py
@@ -3,11 +3,10 @@
 # Distributed under the 3-clause BSD license, see accompanying file LICENSE
 # or https://github.com/scikit-hep/particle for details.
 
-
 from typing import Tuple
 
 from .converters import LHCbName2PDGIDBiMap
-from .utils import from_lhcb_name, to_lhcb_name
+from .functions import from_lhcb_name, to_lhcb_name
 
 __all__ = (
     "from_lhcb_name",
