+++ src/particle/lhcb/converters.py
@@ -7,6 +7,7 @@ from ...converters.bimap import BiMap
 from ...pdgid import PDGID
 from .. import data
 
+
 LHCbName2PDGIDBiMap = BiMap(
     PDGID,
     str,
